Reorganize layout
Prior to this commit, the website was designed in order to only view
pull requests for 4 repositories. If you wanted to display many more
than that, then the "# of requests per repo" graph did not display
nicely.

This commit reorganizes the layout, removing the "# of requests per
repo" graph entirely (until I can find a better way to include it), and
providing a sidebar that lets you filter data for individual pages.