Re-add sorting to pull-requests-per-repo graph
Prior to this commit, I had removed the sorting-by-pull-request-count
feature from the pull-requests-per-repo graph because (as documented in
an earlier commit), that made the bars needle thin for some reason. I
contacted the developer for dc.js, and he had no idea why the frak that
was happening, but he showed me how to force the bars to be the width I
wanted. So this commit re-adds the sort feature.