Add spec tests for web_app.rb
Prior to this commit, there were no spec tests for `web_app.rb`. This
commit adds some (very simple) tests for each of the Sinatra routes.
These tests check 1) that the page can be be reached, 2) where relevant,
some routes return `Not Found` when passed an invalid module name, and
3) where relevant, the route returns a JSON string. More tests to come.