Add tests to JSON routes for web_app
Prior to this commit, there were tests in `web_app_spec.rb` that checked
that the routes that return JSON  did in fact return JSON, and that they
returned either an array or a hash (as applicable). This commit adds
more detailed tests, which check that the routes return hash(es) with
the right keys.